Add min- and max-cut data for instances in current training data
The following min- and max-cut data was generated with
compute_min_max_for_graphs.py on instances that exist in training data
present in the repo. Some training data was "ignored" as there's a bug
in the optimized_qaoa_angles. Additional instances can be computed with
the same script and added to data/minmax_cuts later on.