Add script to fix old TRATrainer data
The old TQATrainer from qaoa-training-pipeline had a bug where the
results dictionary swapped optimized_params and optimized_qaoa_angles.
This commit adds a script that reads a training JSON file and corrects
for this bug. The qaoa_training_pipeline_version was increased to 16.
However, to make sure we keep track of the version with which the
training data was generated, we do not update
qaoa_training_pipeline_version. Instead, we add a new
tqa_trainer_fix_applied entry that has the value `true`. This way we can
keep track of the original qaoa-training-pipeline version and ensure we
don't double-correct TQA results.