Add replace functionality for SummaryTable.add_minmax_cut_data
Previous versions of SummaryTable required recreating instances if
minmax cut data was updated. This was because minmax cut data that
referred to the same graph instance raised an error when loaded. This
prevents multiple JSON files referring to the same graph instance, a
feature we need. However, this prevented reloading minmax cut data
after new files were generated. The new `replace` parameter will
overwrite stored min- and max-cut data in SummaryTable if set to True.
The default, when replace=False, is the old behaviour.
gen_summary_tables.ipynb is not updated as we want the error to be
raised in case of conflicting minmax cut data JSON files.